skills/wechat-draft-publisher/requirements.txt
httpx>=0.27,<1
Markdown>=3.6,<4
Pillow>=10,<12
